more miscellaneous cleanups
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2662 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/files.c
@@ -1697,7 +1697,8 @@ int write_file(const char *name, FILE *f_open, bool tmp, int append,
 	/* Update originalfilestat to reference the file as it is now. */
 	stat(filename, &originalfilestat);
 #endif
-	statusbar(P_("Wrote %lu line", "Wrote %lu lines", lineswritten),
+	statusbar(P_("Wrote %lu line", "Wrote %lu lines",
+		(unsigned long)lineswritten),
 		(unsigned long)lineswritten);
 	UNSET(MODIFIED);
 	titlebar(NULL);
